Fix Utils.kt
